refactor(be): ensure consistency (#9)
* refactor(be): centralize pagination query params dependency

* refactor(schema): unify base schema and pagination response

introduce BaseSchema with camelCase aliases and from_attributes,
replace duplicated Pagination models with generic PaginatedResponse,
and update routers to return typed paginated responses

* refactor(product): store product price as bigint

migrate products.price from numeric/decimal to bigint and update DTOs,
domain entity, ORM model, and pydantic schemas to use int types

* refactor(products): extract keyword and pagination query helpers

* refactor(users): align list filters and reorder deps

ensure consistency

* fix(students): return validated student response in /me

* fix(stores): rename q query param to search

* fix(products): rename keyword query alias to search

* style(be): remove trailing periods from error messages

* feat(users): return updated user on profile and phone changes

update /users/me and phone change endpoints to respond with
UserResponse and 200 status instead of 204

* fix(students): return updated student on PATCH /me

* refactor(stores): split store responses into summary and detail

* refactor(products): return 204 on delete and drop message schema

* fix(users): return 200 status for change phone endpoint

* fix(stores): correct store owner detail response mapping

+++ backend/app/products/dto.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from decimal import Decimal
 
 from app.products.enum import ProductCategory
 from app.stores.entity import Store
@@ -9,7 +8,7 @@ from app.stores.entity import Store
 class CreateProductDTO:
     store: Store
     name: str
-    price: Decimal
+    price: int
     category: ProductCategory
     description: str | None
     photo_url: str | None
